Add _I subtype to operational discovery. (#9754)
* Add _I subtype to operational discovery.

* Fix tv app for new discovery filter type.

* Use BytesToHex rather than PRIX64.

* Doxygen, add comment from review.

+++ examples/platform/linux/ControllerShellCommands.cpp
@@ -65,7 +65,7 @@ static CHIP_ERROR discover(bool printHeader)
         streamer_printf(sout, "Discover:        ");
     }
 
-    Mdns::DiscoveryFilter filter(Mdns::DiscoveryFilterType::kNone, (uint16_t) 0);
+    Mdns::DiscoveryFilter filter(Mdns::DiscoveryFilterType::kNone, (uint64_t) 0);
     gCommissioner->DiscoverCommissionableNodes(filter);
 
     streamer_printf(sout, "done\r\n");
